moving structs to separate files
