+++ examples/simple_aggregate_signature/mod.rs
@@ -1,0 +1,2 @@
+pub(crate) mod signer;
+pub(crate) mod threshold_signature;
